+++ packages/at_client/lib/src/preference/at_client_config.dart
@@ -10,7 +10,7 @@ class AtClientConfig {
 
   /// Represents the at_client version.
   /// Must always be the same as the actual version in pubspec.yaml
-  final String atClientVersion = '3.0.54';
+  final String atClientVersion = '3.0.57';
 
   /// Represents the client commit log compaction time interval
   ///