+++ packages/at_client/lib/src/service/notification_service_impl.dart
@@ -347,7 +347,8 @@ class NotificationServiceImpl
       await atClientValidation.validateNotificationRequest(
           _atClientManager.secondaryAddressFinder!,
           notificationParams,
-          _atClient.getPreferences()!);
+          _atClient.getPreferences()!,
+          _atClient.getCurrentAtSign()!);
       // Get the EncryptionInstance to encrypt the data.
       var atKeyEncryption = atKeyEncryptionManager.get(
           notificationParams.atKey, _atClient.getCurrentAtSign()!);
@@ -358,6 +359,7 @@ class NotificationServiceImpl
               atKeyEncryption)
           .transform(notificationParams);
 
+      notificationValueValidation(notificationParams, builder);
       // Run the notify verb on the remote secondary instance.
       await _atClient.getRemoteSecondary()?.executeVerb(builder);
       if (onSentToSecondary != null) {
@@ -388,6 +390,34 @@ class NotificationServiceImpl
             notificationParams, notificationResult, onSuccess, onError));
         return notificationResult;
       }
+    }
+  }
+
+  /// Checks the size the notification value. If the size exceeds the [AtClientPreference.maxDataSize]
+  /// [BufferOverFlowException] is thrown.
+  ///
+  @visibleForTesting
+  void notificationValueValidation(
+      NotificationParams notificationParams, NotifyVerbBuilder builder) {
+    switch (notificationParams.messageType) {
+      case MessageTypeEnum.key:
+        // Since value is not mandatory in AtNotification, perform validation only if
+        // value is not null.
+        if (builder.value != null &&
+            builder.value.length > _atClient.getPreferences()!.maxDataSize) {
+          throw BufferOverFlowException(
+              'The length of value exceeds the maximum allowed length. Maximum buffer size is ${_atClient.getPreferences()!.maxDataSize} bytes. Found ${builder.value.toString().length} bytes');
+        }
+        break;
+
+      case MessageTypeEnum.text:
+        // When messageType is text, the "text" to notify is added to the key. Hence validating
+        // the key length
+        if (builder.atKey!.length > _atClient.getPreferences()!.maxDataSize) {
+          throw BufferOverFlowException(
+              'The length of value exceeds the maximum allowed length. Maximum buffer size is ${_atClient.getPreferences()!.maxDataSize} bytes. Found ${builder.value.toString().length} bytes');
+        }
+        break;
     }
   }