+++ tests/at_end2end_test/test/bypasscache_test.dart
@@ -1,3 +1,5 @@
+import 'dart:convert';
+
 import 'package:at_client/at_client.dart';
 import 'package:at_end2end_test/config/config_util.dart';
 import 'package:at_end2end_test/src/sync_initializer.dart';
@@ -5,23 +7,55 @@ import 'package:at_end2end_test/src/test_initializers.dart';
 import 'package:at_end2end_test/src/test_preferences.dart';
 
 import 'package:test/test.dart';
+import 'package:version/version.dart';
 
 void main() async {
-  late AtClientManager currentAtClientManager;
-  late AtClientManager sharedWithAtClientManager;
-  late String currentAtSign;
-  late String sharedWithAtSign;
+  late AtClient clientOne;
+  late AtClient clientTwo;
+  late String atSignOne;
+  late String atSignTwo;
   final namespace = 'wavi';
 
   setUpAll(() async {
-    currentAtSign = ConfigUtil.getYaml()['atSign']['firstAtSign'];
-    sharedWithAtSign = ConfigUtil.getYaml()['atSign']['secondAtSign'];
+    atSignOne = ConfigUtil.getYaml()['atSign']['firstAtSign'];
+    atSignTwo = ConfigUtil.getYaml()['atSign']['secondAtSign'];
 
     await TestSuiteInitializer.getInstance()
-        .testInitializer(currentAtSign, namespace);
+        .testInitializer(atSignOne, namespace);
     await TestSuiteInitializer.getInstance()
-        .testInitializer(sharedWithAtSign, namespace);
+        .testInitializer(atSignTwo, namespace);
   });
+
+  Future<Version> getVersion(AtClient atClient, Version ifNull) async {
+    var infoResult = await atClient.getRemoteSecondary()!.executeCommand('info\n');
+
+    expect(infoResult, startsWith('data:{'));
+    infoResult = infoResult!.replaceAll('data:', '');
+    var versionString = jsonDecode(infoResult)['version'];
+    if (versionString == null) {
+      return ifNull;
+    } else {
+      // Since secondary version has gha<number> appended, remove the gha number from version
+      // Hence using split.
+      return Version.parse(versionString.split('+')[0]);
+    }
+  }
+
+  Future<void> setAtSignOneAutoNotify(bool autoNotify) async {
+    //  reset the autoNotify to true
+    clientOne = (await AtClientManager.getInstance().setCurrentAtSign(
+        atSignOne,
+        namespace,
+        TestPreferences.getInstance().getPreference(atSignOne))).atClient;
+
+    var configResult = await clientOne
+        .getRemoteSecondary()!
+        .executeCommand('config:set:autoNotify=$autoNotify\n', auth: true);
+    if (configResult == null) {
+      fail('failed to set auto config to $autoNotify');
+    }
+    expect(configResult.contains('data:ok'), true);
+  }
 
   /// The purpose of this test is to verify the following:
   /// 1. Share a key from atsign_1 to atsign_2 with ttr, with autoNotify:true
@@ -33,93 +67,103 @@ void main() async {
   test('bypass cache test', () async {
     var verificationKey = AtKey()
       ..key = 'verificationNumber'
-      ..sharedWith = sharedWithAtSign
+      ..sharedWith = atSignTwo
       ..metadata = (Metadata()..ttr = 1000);
-    var oldValue = '0873';
-    // updating the key to the currentAtSign
-    currentAtClientManager = await AtClientManager.getInstance()
-        .setCurrentAtSign(currentAtSign, namespace,
-            TestPreferences.getInstance().getPreference(currentAtSign));
+    var initialValue = '0873';
+
+    // Ensure autoNotify is set to true to begin with
+    await setAtSignOneAutoNotify(true);
+
+    // As atSignOne: Put the initial value
+    clientOne = (await AtClientManager.getInstance()
+        .setCurrentAtSign(atSignOne, namespace,
+            TestPreferences.getInstance().getPreference(atSignOne))).atClient;
     var putResult =
-        await currentAtClientManager.atClient.put(verificationKey, oldValue);
+        await clientOne.put(verificationKey, initialValue);
     expect(putResult, true);
     // Sync the data to the remote secondary
     await E2ESyncService.getInstance()
-        .syncData(currentAtClientManager.atClient.syncService);
+        .syncData(clientOne.syncService);
 
-    // Setting sharedWithAtSign atClient instance to context.
-    sharedWithAtClientManager = await AtClientManager.getInstance()
-        .setCurrentAtSign(sharedWithAtSign, namespace,
-            TestPreferences.getInstance().getPreference(sharedWithAtSign));
+    // Give it a couple of seconds to propagate from one atServer to the other
+    await Future.delayed(Duration(seconds: 2));
+
+    // As atSignTwo: do a sync, and do the get, to verify we have the initial value
+    clientTwo = (await AtClientManager.getInstance()
+        .setCurrentAtSign(atSignTwo, namespace,
+            TestPreferences.getInstance().getPreference(atSignTwo))).atClient;
     await E2ESyncService.getInstance()
-        .syncData(sharedWithAtClientManager.atClient.syncService);
+        .syncData(clientTwo.syncService);
     var getKey = AtKey()
       ..key = 'verificationNumber'
-      ..sharedBy = currentAtSign;
-    var getResult = await sharedWithAtClientManager.atClient.get(getKey);
-    expect(getResult.value, oldValue);
+      ..sharedBy = atSignOne;
+    var getResult = await clientTwo.get(getKey);
+    expect(getResult.value, initialValue);
     print('get Result is $getResult');
 
-    // Setting currentAtSign atClient instance to context.
-    await AtClientManager.getInstance().setCurrentAtSign(currentAtSign,
-        namespace, TestPreferences.getInstance().getPreference(currentAtSign));
+    // As atSignOne
+    clientOne = (await AtClientManager.getInstance().setCurrentAtSign(atSignOne,
+        namespace, TestPreferences.getInstance().getPreference(atSignOne))).atClient;
     try {
-      // set auto notify to false
-      var configResult = await currentAtClientManager.atClient
-          .getRemoteSecondary()!
-          .executeCommand('config:set:autoNotify=false\n', auth: true);
-      expect(configResult, contains('data:ok'));
-      // adding a delay for 2 seconds till the config value gets updated
-      await Future.delayed(Duration(seconds: 3));
-      // Updating the same key with a new value
+      // Set autoNotify to false so that the update doesn't propagate to atSignTwo automatically
+      await setAtSignOneAutoNotify(false);
+
+      // Put a new value
       var verificationKeyNew = AtKey()
         ..key = 'verificationNumber'
-        ..sharedWith = sharedWithAtSign;
+        ..sharedWith = atSignTwo;
       var newValue = '9900';
-      var newPutResult = await currentAtClientManager.atClient
+      var newPutResult = await clientOne
           .put(verificationKeyNew, newValue);
       expect(newPutResult, true);
-      await E2ESyncService.getInstance()
-          .syncData(currentAtClientManager.atClient.syncService);
 
-      // Setting sharedWithAtSign atClient instance to context.
-      sharedWithAtClientManager = await AtClientManager.getInstance()
-          .setCurrentAtSign(sharedWithAtSign, namespace,
-              TestPreferences.getInstance().getPreference(sharedWithAtSign));
+      // Sync the data to the remote secondary
       await E2ESyncService.getInstance()
-          .syncData(sharedWithAtClientManager.atClient.syncService);
-      //  get result with bypassCache set to true
-      // should return the newly updated value
-      getResult = await sharedWithAtClientManager.atClient.get(getKey,
+          .syncData(clientOne.syncService);
+
+
+      // As atSignTwo
+      clientTwo = (await AtClientManager.getInstance()
+          .setCurrentAtSign(atSignTwo, namespace,
+              TestPreferences.getInstance().getPreference(atSignTwo))).atClient;
+
+      // Sync - after this we still should have the old value
+      await E2ESyncService.getInstance()
+          .syncData(clientTwo.syncService);
+
+      // get result with bypassCache set to false
+      // should still return the initial value
+      getResult = await clientTwo.get(getKey,
+          getRequestOptions: GetRequestOptions()..bypassCache = false);
+      print('get result with bypass cache false $getResult');
+      expect(getResult.value, initialValue);
+
+      // get result with bypassCache set to true
+      // should return the new value
+      getResult = await clientTwo.get(getKey,
           getRequestOptions: GetRequestOptions()..bypassCache = true);
       print('get result with bypass cache true $getResult');
       expect(getResult.value, newValue);
-      // get Result with byPassCache set to false
-      // should return the old value
-      // adding a delay
-      await Future.delayed(Duration(seconds: 3));
-      var getResultWithFalse = await sharedWithAtClientManager.atClient.get(
-          getKey,
-          getRequestOptions: GetRequestOptions()..bypassCache = false);
-      print('get result with bypass cache false $getResultWithFalse');
-      expect(getResultWithFalse.value, oldValue);
-      //  reset the autoNotify to false
-      await AtClientManager.getInstance().setCurrentAtSign(
-          currentAtSign,
-          namespace,
-          TestPreferences.getInstance().getPreference(currentAtSign));
-    } finally {
-      // Moving the config:set:autoNotify=true to finally block because if
-      // test fails in between the autoNotify on the atSign remains false.
-      var configResult = await currentAtClientManager.atClient
-          .getRemoteSecondary()!
-          .executeCommand('config:set:autoNotify=true\n', auth: true);
-      if (configResult == null) {
-        assert(fail('failed to set auto config to true'));
+
+      // Sync - after this we should now have the new value
+      await E2ESyncService.getInstance()
+          .syncData(clientTwo.syncService);
+
+      // get Result with byPassCache set to false again
+      // should also now return the new value, since cached value will have been updated with the
+      // results of the remote lookup, and the cached value will have been synced to the client
+      // Note: This will only work on server versions which have the fix, from version 3.0.29 onwards
+      Version serverTwoVersion = await getVersion(clientTwo, Version(3, 0, 29));
+      if (serverTwoVersion >= Version(3, 0, 29)) {
+        var getResultWithFalse = await clientTwo.get(
+            getKey,
+            getRequestOptions: GetRequestOptions()
+              ..bypassCache = false);
+        print('get result with bypass cache false $getResultWithFalse');
+        expect(getResultWithFalse.value, newValue);
       }
-      assert(configResult!.contains('data:ok'), true);
+    } finally {
+      await setAtSignOneAutoNotify(true);
     }
-    //Setting the timeout to prevent termination of test, since we have Future.delayed
-    // for 30 Seconds.
-  }, timeout: Timeout(Duration(minutes: 5)));
+  });
 }